% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_tilematrix.R
\name{gg_tilematrix.matrix}
\alias{gg_tilematrix.matrix}
\title{Plot a matrix as a ggplot2 object}
\usage{
gg_tilematrix.matrix(
  .matrix,
  .geom = list(geom_tile(color = "pink", fill = "blue"), geom_text()),
  .scale_coord = list(scale_y_discrete(limits = base::rev), scale_x_discrete(position =
    "top"), coord_fixed()),
  .theme = list(theme_bw()),
  .layers = list(labs(x = NULL, y = NULL))
)
}
\arguments{
\item{.matrix}{A matrix to plot}

\item{.layers}{}
}
\description{
Plot a matrix as a ggplot2 object
}
\examples{
library(ggplot2)
matrices$unnamed |>
  gg_tilematrix.matrix()

}
