% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triples_to_matrix.R, R/zzz_aliases.R
\name{matrix_to_triples}
\alias{matrix_to_triples}
\alias{triples_from_matrix}
\title{Pivots a matrix into (bi-graph) triples}
\usage{
matrix_to_triples(
  matrix,
  x_names_to = "x_name",
  y_names_to = "y_name",
  values_to = "value",
  drop_na = TRUE
)

triples_from_matrix(
  matrix,
  x_names_to = "x_name",
  y_names_to = "y_name",
  values_to = "value",
  drop_na = TRUE
)
}
\arguments{
\item{matrix}{A matrix to pivot}

\item{x_names_to}{A string specifying the new column for
the names of the first/x-dimension of \code{matrix}}

\item{y_names_to}{A string specifying the new column for
the names of the second/y-dimension  \code{matrix}}

\item{values_to}{A character vector specifying the new column to create
from the cell values of \code{matrix}}

\item{drop_na}{Boolean option to drop \code{x}-\code{y} pairs with missing values.}
}
\description{
Converts a named (adjacency) matrix into (graph) triples,
dropping any NA cells by default.
}
\details{
Uses \link[tidyr]{pivot_longer} and reverses \code{\link[=triples_to_matrix]{triples_to_matrix()}}.
Setting \code{drop_na=FALSE} retain the complete set of graph edges.
If \code{matrix} is unnamed, default names are generated by \code{\link[base:as.data.frame]{base::as.data.frame()}}
}
\examples{

## named matrices
blx_mtx <- triples$blx_tbl |>
  triples_to_matrix(from, to, weighted)
return_triples <- blx_mtx |>
  triples_from_matrix("from", "to", "weighted")

identical(triples$blx_tbl, return_triples)

## Setting `drop_na=FALSE` retains the `x`-`y` pairs with missing values,
return_with_na <- blx_mtx |>
  triples_from_matrix("from", "to", "weighted", drop_na = FALSE)

## which is equivalent to using `tidyr::complete()` on the original triples
complete_triples <- triples$blx_tbl |>
  tidyr::complete(from, to)

identical(return_with_na, complete_triples)

## names are generated for unnamed matrices
matrices$unnamed |>
  matrix_to_triples(drop_na = FALSE)
}
