% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triples_to_matrix.R, R/zzz_aliases.R
\name{triples_to_matrix}
\alias{triples_to_matrix}
\alias{matrix_from_tripes}
\title{Pivot (bi-graph) triples into (bi-adjacency) matrix}
\usage{
triples_to_matrix(
  triples,
  x_names_from = x_name,
  y_names_from = y_name,
  values_from = value,
  values_fill = NA
)

matrix_from_tripes(
  triples,
  x_names_from = x_name,
  y_names_from = y_name,
  values_from = value,
  values_fill = NA
)
}
\arguments{
\item{triples}{A data frame to pivot. Unused columns are dropped.}

\item{x_names_from}{column to get matrix row row names from.}

\item{y_names_from}{column to get matrix column names from.}

\item{values_from}{column to get cell values from.}

\item{values_fill}{Optionally, a (scalar) value that specifies what each
\code{value} should be filled in with when missing.

This can be a named list if you want to apply different fill values to
different value columns.}
}
\description{
Converts a data frame of triples (\code{x_names}, \code{y_names}, \code{values}) into
a matrix with \code{x_names} and \code{y_names} respectively
for row and column names and cell values from \code{values}.
}
\details{
The triples define graph edges between \code{x_names} and \code{y_names} with
attribute \code{value}. \code{x_names} and \code{y_names} are treated as disjoint nodes sets
of the bi-partite graph,
and the resultant matrix corresponds to the bi-adjacency matrix.

Uses \link[tidyr]{pivot_wider} and makes explicit
all possible pairs of \code{x_names} and \code{y_names}.
}
\examples{
triples$blx_tbl

triples$blx_tbl |>
  triples_to_matrix(from, to, weighted)
}
